Refactor controller for symbols upload
