+++ src/NuGetGallery/Strings.resx
@@ -157,7 +157,7 @@
     <value>A package file is required.</value>
   </data>
   <data name="UploadFileMustBeNuGetPackage" xml:space="preserve">
-    <value>The file must be a .nupkg file.</value>
+    <value>The file must be a .nupkg or .snupkg file.</value>
   </data>
   <data name="SuccessfullyUploadedPackage" xml:space="preserve">
     <value>You successfully uploaded {0} {1}.</value>
