+++ src/NuGetGallery/Views/Packages/UploadPackage.cshtml
@@ -30,7 +30,7 @@
                             <input type="text" class="form-control" id="file-select-feedback" value="Browse or Drop files to select a package..." aria-label="Upload Your Package" readonly />
                             <label for="input-select-file" id="PackageFileLabel" class="input-group-btn">
                                 <span id="browse-for-package-button" class="btn btn-primary form-control" tabindex="0" aria-label="Browse for package" role="button">
-                                    Browse&hellip;<input type="file" name="UploadFile" accept=".nupkg" aria-labelledby="PackageFileLabel" id="input-select-file" style="display:none;" />
+                                    Browse&hellip;<input type="file" name="UploadFile" accept=".nupkg,.snupkg" aria-labelledby="PackageFileLabel" id="input-select-file" style="display:none;" />
                                 </span>
                             </label>
                         </div>
